Add a legend to the map
Add a legend to the map. Add the LRT stations and lines into a group
(lrt) and remove their individual addTo(map) commands. Add the Nice Ride
stations and Light Rail to the legend. Set the legend to collapse if the
screen is under 767 pixels (tablets/phones/small browsers). Add the
legend to the map.